adding bot names to 1:1 chat titles
