+++ packages/mgt-chat/src/components/ChatHeader/OneToOneChatHeader.tsx
@@ -3,6 +3,8 @@ import { AadUserConversationMember, Chat } from '@microsoft/microsoft-graph-type
 import { Person } from '@microsoft/mgt-react';
 import { ChatHeaderProps } from './ChatTitle';
 import { makeStyles } from '@fluentui/react-components';
+import { useBotInfo } from '../../statefulClient/useBotInfo';
+import { BotAvatar } from '../ChatAvatar/BotAvatar';
 
 const getOtherParticipantUserId = (chat?: Chat, currentUserId = '') =>
   (chat?.members as AadUserConversationMember[])?.find(m => m.userId !== currentUserId)?.userId;
@@ -14,16 +16,36 @@ const useStyles = makeStyles({
   }
 });
 export const OneToOneChatHeader = ({ chat, currentUserId }: ChatHeaderProps) => {
+  const botInfo = useBotInfo();
   const styles = useStyles();
   const id = getOtherParticipantUserId(chat, currentUserId);
-  return id ? (
-    <Person
-      className={styles.person}
-      userId={id}
-      view="oneline"
-      avatarSize="small"
-      personCardInteraction="hover"
-      showPresence={true}
-    />
-  ) : null;
+  if (!chat?.id) return null;
+  if (id) {
+    return (
+      <Person
+        className={styles.person}
+        userId={id}
+        view="oneline"
+        avatarSize="small"
+        personCardInteraction="hover"
+        showPresence={true}
+      />
+    );
+  } else if (botInfo?.chatBots.has(chat.id)) {
+    return (
+      <>
+        {Array.from(botInfo.chatBots.get(chat.id)!).map(bot =>
+          bot.teamsAppDefinition?.bot?.id ? (
+            <BotAvatar
+              key={bot.teamsAppDefinition.bot.id}
+              chatId={chat.id!}
+              avatarId={bot.teamsAppDefinition.bot.id}
+              view="oneline"
+            />
+          ) : null
+        )}
+      </>
+    );
+  }
+  return null;
 };
